requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.0.0
numpy==1.23.5
scikit-learn==1.3.0
yfinance==0.2.28
alpha-vantage==2.3.1
tensorflow==2.13.0
requests==2.31.0
nltk==3.8.1
matplotlib==3.7.0


